Move a lot of code back into ChartView.
